requirements.txt
Django==1.10
django-pure-pagination==0.3.0
Markdown==2.6.8
mysqlclient==1.3.10
olefile==0.44
Pillow==4.1.1
Pygments==2.2.0
